+++ api/managementpb/ia/json/client/templates/list_templates_responses.go
@@ -994,19 +994,19 @@ type TemplatesItems0ParamsItems0Float struct {
 	HasDefault bool `json:"has_default,omitempty"`
 
 	// Default value if has_default is true.
-	Default float32 `json:"default,omitempty"`
+	Default float64 `json:"default,omitempty"`
 
 	// True if minimal valid value is set.
 	HasMin bool `json:"has_min,omitempty"`
 
 	// Minimal valid value (inclusive) if has_min is true.
-	Min float32 `json:"min,omitempty"`
+	Min float64 `json:"min,omitempty"`
 
 	// True if maximal valid value is set.
 	HasMax bool `json:"has_max,omitempty"`
 
 	// Maximal valid value (inclusive) if has_max is true.
-	Max float32 `json:"max,omitempty"`
+	Max float64 `json:"max,omitempty"`
 }
 
 // Validate validates this templates items0 params items0 float
